Update index.php
The data is requested through the API from another .php file thourgh JS
-----------------------------------------------------------------------------------------
Display the downloaded data in a table showing task, title, description and colorCode. The displayed HTML element for the colorCode should have its color set accordingly
Create a search which allows searching for any of the data in the table
Implement auto-refresh functionality which requests the data from above every 60 minutes and updates the table with the new data without reloading the web page
Outside the table, create a button that opens a modal. In this modal, there should be another button that allows you to select any image from the file system. When you have selected the image, it will be displayed in the modal